requirements.txt
bullet==2.2.0
click==8.1.3
MouseInfo==0.1.3
Pillow==9.1.1
pycryptodomex==3.15.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
typer==0.6.1
